VMS-30883: Switch QnTimePeriodList to std implementation
(cherry picked from commit 47582bc277a402c441bfb4aee4de69d7b7a43246)
(cherry picked from commit 8b5a3fa1bb911d87bf89821c6ae9f820fb3a0e68)
(cherry picked from commit 4a25b7982842c17bf4a2eded7e8792bc9c8f954c)

+++ vms/client/nx_vms_client_core/src/camera/loaders/flat_camera_data_loader.cpp
@@ -35,7 +35,7 @@ QString filterRepresentation(const QString& filter, Qn::TimePeriodContent dataTy
 QString periodsLogString(const QnTimePeriodList& periodList, int count = 100)
 {
     QStringList result;
-    int index = std::max(periodList.size() - count, 0);
+    int index = std::max((int) periodList.size() - count, 0);
     for (int i = index; i < periodList.size(); ++i)
         result.append(
             nx::format(
